+++ smolseg/models/builder.py
@@ -1,15 +1,23 @@
 import torch.nn as nn
+from smolseg.models.erfnet import ERFNet
 from smolseg.models.smp_models import SmpUnetWrapper
 from smolseg.models.torchvision_models import TorchvisionModelWrapper
 
 def build_model(
     model_name: str,
     num_classes: int,
-    encoder_name: str = "resnet50",
     pretrained: bool = False,
 ) -> nn.Module:
     
-    if model_name in ["unet_seresnext26d"]:
+    if model_name in [
+        "unet_seresnext26d",
+        "deeplabv3plus_seresnext26d",
+        "deeplabv3plus_regnetz_b16"
+    ]:
+        if model_name.endswith("seresnext26d"):
+            encoder_name = "tu-seresnext26d_32x4d"
+        elif model_name.endswith("regnetz_b16"):
+            encoder_name = "tu-regnetz_b16"
         return SmpUnetWrapper(
             model_name=model_name,
             num_classes=num_classes,
@@ -22,5 +30,7 @@ def build_model(
             num_classes=num_classes,
             pretrained=pretrained,
         )
+    elif model_name in ["erfnet"]:
+        return ERFNet(num_classes=num_classes, in_channels=3)
     else:
         raise NotImplementedError(f"Model {model_name} not implemented")