+++ smolseg/models/smp_models.py
@@ -5,9 +5,15 @@ class SmpUnetWrapper(nn.Module):
 
     def __init__(self, model_name: str, num_classes: int, encoder_name: str = "resnet50", encoder_weights: str = "imagenet"):
         super().__init__()
-        if model_name == "unet_seresnext26d":
+        if model_name.startswith("unet"):
             self.model = smp.Unet(
-                encoder_name="tu-seresnext26d_32x4d",
+                encoder_name=encoder_name,
+                encoder_weights=encoder_weights,
+                classes=num_classes
+            )
+        elif model_name.startswith("deeplabv3plus"):
+            self.model = smp.DeepLabV3Plus(
+                encoder_name=encoder_name,
                 encoder_weights=encoder_weights,
                 classes=num_classes
             )